+++ lib/spicerack/spec_helper.rb
@@ -2,9 +2,6 @@
 
 require "rspice"
 require "redis_hash/spec_helper"
+require "substance/rspec"
 require "technologic/spec_helper"
 require "tablesalt/spec_helper"
-
-require_relative "rspec/custom_matchers"
-require_relative "rspec/shared_examples"
-require_relative "rspec/shoulda_matcher_helper"