+++ substance/lib/substance/rspec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "rspec/custom_matchers"
+require_relative "rspec/shared_examples"
+require_relative "rspec/shoulda_matcher_helper"