+++ substance/lib/substance/rspec/custom_matchers.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "custom_matchers/define_argument"
+require_relative "custom_matchers/define_attribute"
+require_relative "custom_matchers/define_option"